requirements.txt
Django>=5.0,<5.1
django-environ>=0.11
django-filter>=24.3
djangorestframework>=3.15
Pillow>=10.3

